Lighting sequence complete. Lots of magic numbers

+++ SunriseAlarmClock/lights.h
@@ -3,13 +3,14 @@
 
 #include "types.h"
 #include <avr/io.h>
+#include <math.h>
 
 class Lights {
     public:
     static void enable();
     static void setWhite(unsigned int intensity);
     static void setRed(unsigned int intensity);
-    static void setCombined(unsigned int intensity);
+    static void setCombined(unsigned long intensity);
 };
 
 #endif /* LIGHTS_H_ */
